requirements.txt
psutil==5.9.0
pyserial==3.5
